requirements.txt
kivy==2.3.1
kivymd==1.2.0
pymongo
Pillow
langchain
langchain-community
faiss-cpu

